optimize product provider
